src/SSDTLifecycleExtension/MVVM/AsyncCommand.cs
namespace SSDTLifecycleExtension.MVVM
{
    using System;
    using System.Threading.Tasks;
    using System.Windows.Input;
    using JetBrains.Annotations;

    public class AsyncCommand : IAsyncCommand
    {
        public event EventHandler CanExecuteChanged;

        private bool _isExecuting;
        private readonly Func<Task> _execute;
        private readonly Func<bool> _canExecute;
        private readonly IErrorHandler _errorHandler;

        public AsyncCommand([NotNull] Func<Task> execute,
                            [NotNull] Func<bool> canExecute,
                            [NotNull] IErrorHandler errorHandler)
        {
            _execute = execute ?? throw new ArgumentNullException(nameof(execute));
            _canExecute = canExecute ?? throw new ArgumentNullException(nameof(canExecute));
            _errorHandler = errorHandler ?? throw new ArgumentNullException(nameof(errorHandler));
        }

        public bool CanExecute()
        {
            return !_isExecuting && _canExecute.Invoke();
        }

        public async Task ExecuteAsync()
        {
            if (CanExecute())
            {
                try
                {
                    _isExecuting = true;
                    RaiseCanExecuteChanged();
                    await _execute();
                }
                finally
                {
                    _isExecuting = false;
                }
            }

            RaiseCanExecuteChanged();
        }

        public void RaiseCanExecuteChanged()
        {
            CanExecuteChanged?.Invoke(this, EventArgs.Empty);
        }

        #region Explicit ICommand implementations

        bool ICommand.CanExecute(object parameter)
        {
            return CanExecute();
        }

        void ICommand.Execute(object parameter)
        {
            ExecuteAsync().FireAndForget(this, _errorHandler);
        }

        #endregion
    }
}